[react-dom] Update types for v18
